requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.4.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==21.12b0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
Deprecated==1.2.13
distlib==0.3.6
Django==4.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-storages==1.13.2
djangorestframework==3.12.4
dnspython==2.1.0
fake-useragent==0.1.11
filelock==3.6.0
fingerprint-pro-server-api-sdk==2.1.0
google-api-core==2.11.0
google-api-python-client==2.33.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-cloud-tasks==2.13.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
identify==2.5.22
idna==3.3
importlib-metadata==4.8.2
kombu==5.2.4
lxml==4.6.4
mailchimp-marketing==3.0.72
Markdown==3.3.6
mixpanel==4.9.0
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.0
oauthlib==3.1.1
packaging==21.3
pandas==1.4.3
parse==1.19.0
pathspec==0.9.0
platformdirs==2.4.0
pre-commit==3.2.2
prompt-toolkit==3.0.29
proto-plus==1.22.2
protobuf==4.22.3
psycopg2==2.9.2
py3-validate-email==1.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
pyparsing==3.0.6
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
python-http-client==3.3.4
pytz==2021.3
PyYAML==6.0
redis==4.2.2
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
rsa==4.8
sendgrid==6.9.3
six==1.16.0
SkPy==0.10.4
soupsieve==2.3.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
timeout-decorator==0.5.0
tomli==1.2.2
tqdm==4.62.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.21.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==9.1
whitenoise==6.4.0
whois==0.9.13
wrapt==1.14.0
zerobouncesdk==0.0.6
zipp==3.6.0
